Separate elm-tooling.json handling
This separates the elm-tooling.json handling, preparing us to load more values.
Should also allow starting to cache these calls at some point.
