+++ test/jest.setup.ts
@@ -7,6 +7,9 @@ import { DocumentEvents } from "../src/util/documentEvents";
 
 container.register("Connection", { useValue: mockDeep<Connection>() });
 container.register("ElmWorkspaces", { useValue: [] });
+container.register("ElmToolingJsonManager", {
+  useValue: {},
+});
 container.register("Settings", {
   useValue: new Settings({} as any, {}),
 });
